- fixed 64bits compiling: pointers are on 64 and they need to be cast to interger via long type!
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@3931 689a6050-402a-0410-94f2-e92a70836424

+++ mod_fix.c
@@ -160,7 +160,7 @@ int fixup_uint(void** param)
 	if(str2int(&s, &ui)==0)
 	{
 		pkg_free(*param);
-		*param=(void *)ui;
+		*param=(void *)(unsigned long)ui;
 		return 0;
 	}
 	LM_ERR("bad number <%s>\n", (char *)(*param));
